+++ packages/plugin-file-upload/web/src/icons/xlsIcon.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable max-len */
 import React from 'react';
 
-const xlsIcon = props => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="40" height="42" viewBox="0 0 40 42" {...props}>
+const xlsIcon = ({ styles, ...rest }) => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="40" height="42" viewBox="0 0 40 42" {...rest}>
     <g fill="none" fillRule="evenodd">
       <path fill="#FFF" d="M0 0H4823V3877H0z" transform="translate(-1489 -3614)" />
       <path d="M1489 3613H1529V3655H1489z" transform="translate(-1489 -3614)" />
@@ -12,7 +12,7 @@ const xlsIcon = props => (
           d="M13.63 0.89L7.93 8.34 4.88 4.75 0.5 9.89 6.75 9.89 9.25 9.89 20.5 9.89z"
           transform="translate(-1489 -3614) translate(1492 3613) translate(6 21)"
         />
-        <g stroke="#FFF" strokeLinejoin="bevel">
+        <g className={styles.file_upload_icon_frame} strokeLinejoin="bevel">
           <path
             fill="currentColor"
             d="M32.5 41.5L0.5 41.5 0.5 0.5 19.95 0.5 32.5 13.31z"
@@ -24,7 +24,7 @@ const xlsIcon = props => (
           />
         </g>
         <path
-          fill="#FFF"
+          className={styles.file_upload_icon_background}
           fillRule="nonzero"
           d="M3.29 11.55L0.62 11.55 4.16 5.95 0.8 0.55 3.54 0.55 5.5 4.13 7.52 0.55 10.17 0.55 6.81 5.86 10.38 11.55 7.59 11.55 5.5 7.8z"
           transform="translate(-1489 -3614) translate(1492 3613) translate(11 19)"